break out a types file to hold all of our interfaces
